refactoring: splitting, reusable hooks, structures (#2077)
* refactoring: splitting, reusable hooks, structures

* erros for rewards

* custom integrations memo

---------
