+++ wondrous-bot-admin/src/components/AddFormEntity/components/OptionRewards.tsx
@@ -1,9 +1,9 @@
 import AddIcon from "@mui/icons-material/Add";
 import { Box, Grid } from "@mui/material";
-import { PAYMENT_OPTIONS } from "components/CreateTemplate/RewardUtils";
 import { PoapImage } from "components/CreateTemplate/styles";
 import DeleteIcon from "components/Icons/Delete";
 import { DiscordRoleIcon, NFTIcon, PoapIcon, StoreItemRewardIcon, TokensIcon } from "components/Icons/Rewards";
+import { PAYMENT_OPTIONS } from "components/Rewards/constants";
 import { ButtonIconWrapper } from "components/Shared/styles";
 import { TextLabel } from "components/ViewQuest/styles";
 
@@ -12,14 +12,14 @@ const ICONS_MAP = {
   [PAYMENT_OPTIONS.POAP]: PoapIcon,
   [PAYMENT_OPTIONS.TOKEN]: TokensIcon,
   [PAYMENT_OPTIONS.COMMUNITY_BADGE]: NFTIcon,
-  [PAYMENT_OPTIONS.CMTY_STORE_ITEM]: StoreItemRewardIcon
+  [PAYMENT_OPTIONS.CMTY_STORE_ITEM]: StoreItemRewardIcon,
 };
 
 const RewardContent = ({ reward }) => {
   const Icon = reward?.paymentMethod?.nftMetadata?.mediaUrl
     ? () => <PoapImage src={reward?.paymentMethod?.nftMetadata?.mediaUrl} />
     : ICONS_MAP[reward?.type];
-  
+
   const label = {
     [PAYMENT_OPTIONS.DISCORD_ROLE]: `Role: ${reward?.discordRewardData?.discordRoleName}`,
     [PAYMENT_OPTIONS.TOKEN]: `Token: ${reward?.amount} ${reward?.paymentMethod?.name}`,
