+++ wondrous-bot-admin/src/components/CreateTemplate/MaxInput.tsx
@@ -1,29 +1,9 @@
 import { Box } from "@mui/material";
 import { CustomTextField } from "components/AddFormEntity/components/styles";
 import Switch from "components/Shared/Switch";
-import Tooltip from "@mui/material/Tooltip";
-import InformationTooltip from "components/Icons/information.svg";
-import { StyledInformationTooltip } from "components/Shared/Tooltip";
-import { useMemo } from "react";
-
-const KEYS = {
-  MAX_SUBMISSION: "maxSubmission",
-  MAX_APPROVAL: "maxApproval",
-  MAX_PURCHASE: "maxPurchase",
-  REQUIRE_REVIEW: "requireReview",
-};
-
-const INFO_LABELS = {
-  [KEYS.MAX_SUBMISSION]: "The maximum number of approved submissions each user can submit for this quest",
-  [KEYS.MAX_APPROVAL]: "The total number of approved submissions allowed for this quest",
-  [KEYS.MAX_PURCHASE]: "The maximum number of times a user can purchase this item",
-  [KEYS.REQUIRE_REVIEW]: "If enabled, all submissions will require review before being approved",
-};
 
 const MaxInput = (props) => {
-  const { keyValue, stateKey, handleValueChange, value } = props;
-
-  const infoLabel = INFO_LABELS[stateKey];
+  const { keyValue, handleValueChange, value } = props;
 
   const isActive = value || keyValue?.trim() === "";
 
@@ -46,15 +26,6 @@ const MaxInput = (props) => {
         />
       </Box>
       <Switch {...props} value={isActive} />
-      {/* {infoLabel ? (
-        <StyledInformationTooltip placement="right" title={infoLabel}>
-          <img
-            src={InformationTooltip}
-            alt="information"
-            style={{ width: "16px", height: "16px", marginLeft: "8px" }}
-          />
-        </StyledInformationTooltip>
-      ) : null} */}
     </Box>
   );
 };
