+++ wondrous-bot-admin/src/components/CreateTemplate/index.tsx
@@ -1,32 +1,28 @@
-import { Box, CircularProgress, Divider, Grid, Typography } from "@mui/material";
-import { ErrorText, RoundedSecondaryButton, SharedSecondaryButton } from "components/Shared/styles";
+import { Box, Grid, Typography } from "@mui/material";
+import { RoundedSecondaryButton, SharedSecondaryButton } from "components/Shared/styles";
 import { useContext, useEffect, useMemo, useRef, useState } from "react";
 import AddIcon from "@mui/icons-material/Add";
 import { CampaignOverviewHeader, CampaignOverview } from "./CampaignOverview";
 import PanelComponent from "./PanelComponent";
 import { Panel } from "./styles";
 import AddFormEntity from "components/AddFormEntity";
-import { APEIRON_TYPES, BG_TYPES, QUEST_STATUSES, TYPES } from "utils/constants";
-import { RewardComponent } from "./RewardComponent";
+import { BG_TYPES, QUEST_STATUSES, TYPES } from "utils/constants";
 import PageWrapper from "components/Shared/PageWrapper";
 import Modal from "components/Shared/Modal";
 import { useMutation } from "@apollo/client";
 import { ATTACH_QUEST_STEPS_MEDIA, CREATE_QUEST, REMOVE_QUEST_STEP_MEDIA, UPDATE_QUEST } from "graphql/mutations";
 import GlobalContext from "utils/context/GlobalContext";
 import { useNavigate } from "react-router";
-import { questValidator, ValidationError } from "services/validators";
-import { getPathArray, toCent } from "utils/common";
-import { set } from "lodash";
-import { handleMediaUpload, transformAndUploadMedia } from "utils/media";
+import { questValidator } from "services/validators";
+import { handleMediaUpload } from "utils/media";
 import CreateQuestContext from "utils/context/CreateQuestContext";
-import { PAYMENT_OPTIONS } from "./RewardUtils";
 import { transformQuestConfig } from "utils/transformQuestConfig";
 import useAlerts from "utils/hooks";
-import { DEFAULT_QUEST_SETTINGS_STATE_VALUE, mapAnswersToOptions, reduceConditionalRewards } from "./utils";
+import { DEFAULT_QUEST_SETTINGS_STATE_VALUE } from "./utils";
 import { useTour } from "@reactour/tour";
-import { Warning, WarningAmberOutlined } from "@mui/icons-material";
-import WarningIcon from "components/Icons/WarningIcon";
 import ErrorField from "components/Shared/ErrorField";
+import QuestRewardComponent from "components/Rewards/QuestRewardComponent";
+import { constructRequestBody, handleSaveError, processSave } from "./helpers";
 
 const CreateTemplate = ({
   setRefValue,
@@ -57,7 +53,7 @@ const CreateTemplate = ({
     }
   }, [getQuestById]);
 
-  const { isOpen, setCurrentStep, currentStep, setSteps: setTourSteps, steps: tourSteps } = useTour();
+  const { isOpen, setCurrentStep, setSteps: setTourSteps, steps: tourSteps } = useTour();
 
   const [removeQuestStepMedia] = useMutation(REMOVE_QUEST_STEP_MEDIA);
 
@@ -142,12 +138,6 @@ const CreateTemplate = ({
     setTourSteps(newSteps);
   }, [isOpen]);
 
-  const postRewardsToggle = () => {
-    if (isOpen) {
-      setCurrentStep((prev) => prev + 1);
-    }
-  };
-
   const handleUpdateQuestStepsMedia = async (questId, questSteps, localSteps) => {
     const stepsMedia = await Promise.all(
       localSteps.map(async (step, idx) => {
@@ -210,243 +200,28 @@ const CreateTemplate = ({
     if (Object.keys(errors).length > 0) {
       setErrors({});
     }
-    const {
-      questConditions,
-      requireReview,
-      maxSubmission,
-      isActive,
-      maxApproval,
-      startAt,
-      endAt,
-      level,
-      timeBound,
-      isOnboarding,
-      title,
-      description,
-      conditionLogic,
-    } = questSettings;
-    const filteredQuestConditions = questConditions?.filter((condition) => condition.type && condition.conditionData);
-
-    const body = {
-      title,
-      description,
-      orgId: activeOrg.id,
-      isOnboarding,
-      requireReview: !!requireReview,
-      maxSubmission: maxSubmission ? parseInt(maxSubmission, 10) : null,
-      maxApproval: maxApproval ? parseInt(maxApproval, 10) : null,
-      conditionLogic,
-      category: questSettings.category || null,
-      questConditions: filteredQuestConditions,
-      status: status || (isActive ? QUEST_STATUSES.OPEN : QUEST_STATUSES.INACTIVE),
-      startAt: startAt && timeBound ? startAt.utcOffset(0).startOf("day").toISOString() : null,
-      endAt: endAt && timeBound ? endAt.utcOffset(0).endOf("day").toISOString() : null,
-      pointReward: questSettings.rewards[0].value,
-      submissionCooldownPeriod: questSettings?.submissionCooldownPeriod,
-      level: level ? parseInt(level, 10) : 1,
-      // TODO: refactor this
-      rewards: questSettings.rewards
-        ?.map((reward: any) => {
-          if (reward?.type === PAYMENT_OPTIONS.DISCORD_ROLE) {
-            return {
-              discordRewardData: {
-                discordRoleId: reward?.discordRewardData?.discordRoleId,
-                discordGuildId: reward?.discordRewardData?.discordGuildId,
-                discordRoleName: reward?.discordRewardData?.discordRoleName,
-              },
-              type: reward?.type,
-            };
-          } else if (reward?.type === PAYMENT_OPTIONS.TOKEN || reward?.type === PAYMENT_OPTIONS.COMMUNITY_BADGE) {
-            return {
-              type: PAYMENT_OPTIONS.TOKEN,
-              paymentMethodId: reward?.paymentMethodId,
-              amount: Number(reward?.amount),
-            };
-          } else if (reward?.type === PAYMENT_OPTIONS.POAP) {
-            const { __typename, ...rewardData } = reward?.poapRewardData;
-            return {
-              type: reward?.type,
-              poapRewardData: rewardData,
-            };
-          } else if (reward?.type === PAYMENT_OPTIONS.CMTY_STORE_ITEM) {
-            return {
-              type: reward?.type,
-              storeItemId: reward?.storeItem?.id,
-            };
-          }
-        })
-        ?.filter((reward) => reward),
-      steps: steps.reduce((acc, next, index) => {
-        const step: any = {
-          id: next?._id,
-          type: next.type,
-          order: index + 1,
-          mediaUploads: [],
-          required: next.required === false ? false : true,
-          prompt: next.value?.question || next?.value?.prompt || next.value || null,
-        };
-        const isQuiz = [TYPES.MULTI_QUIZ, TYPES.SINGLE_QUIZ].includes(next.type);
-        if (isQuiz) {
-          step.type = next.value.multiSelectValue;
-          step.options = next.value?.answers
-            ? mapAnswersToOptions(next.value.answers, next.value.withCorrectAnswers)
-            : [];
-          const conditionalRewards = next.value?.answers?.reduce(reduceConditionalRewards, []);
-          if (conditionalRewards.length) {
-            step.conditionalRewards = conditionalRewards;
-          }
-          step.prompt = next.value.question;
-        } else if ([TYPES.LIKE_TWEET, TYPES.RETWEET, TYPES.REPLY_TWEET].includes(next.type)) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            tweetLink: next.value?.tweetLink,
-          };
-        } else if (next.type === TYPES.FOLLOW_TWITTER) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            tweetHandle: next.value?.tweetHandle,
-          };
-        } else if (next.type === TYPES.TWEET_WITH_PHRASE) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            tweetPhrase: next.value?.tweetPhrase,
-          };
-        } else if (next.type === TYPES.LIKE_YT_VIDEO) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            ytVideoLink: next.value?.ytVideoLink,
-          };
-        } else if (next.type === TYPES.SUBSCRIBE_YT_CHANNEL) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            ytChannelLink: next.value?.ytChannelLink,
-          };
-        } else if (next.type === TYPES.LINK_CLICK) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            linkClickUrl: next.value?.linkClickUrl,
-          };
-        } else if (next.type === TYPES.SNAPSHOT_PROPOSAL_VOTE) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            snapshotProposalLink: next.value?.snapshotProposalLink,
-          };
-        } else if (next.type === TYPES.SNAPSHOT_SPACE_VOTE) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            snapshotSpaceLink: next.value?.snapshotSpaceLink,
-            snapshotVoteTimes: Number(next.value?.snapshotVoteTimes),
-          };
-        } else if (next.type === TYPES.DISCORD_MESSAGE_IN_CHANNEL) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            discordMessageType: next.value?.discordMessageType,
-            discordChannelId: next.value?.discordChannelId,
-          };
-        } else if (next.type === TYPES.DISCORD_EVENT_ATTENDANCE) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            discordEventId: next.value?.discordEventId,
-            minDuration: next.value?.minDuration,
-          };
-        } else if (next.type === TYPES.VERIFY_TOKEN_HOLDING) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            tokenAddress: next.value?.verifyTokenAddress,
-            tokenSymbol: next.value?.verifyTokenSymbol,
-            tokenLogoUrl: next.value?.verifyTokenLogoUrl,
-            tokenDecimals: next.value?.verifyTokenDecimals,
-            tokenChain: next.value?.verifyTokenChain,
-            tokenAmount: next.value?.verifyTokenAmount,
-            tokenType: next.value?.verifyTokenType,
-            tokenId: next.value?.verifyTokenId,
-            tokenName: next.value?.verifyTokenName,
-          };
-        } else if (next.type === TYPES.DATA_COLLECTION) {
-          step.prompt = next.value?.prompt;
-          step.options = next?.value?.options
-            ? next.value.options.map((option, idx) => ({
-                position: idx,
-                text: option,
-              }))
-            : null;
-          step["additionalData"] = {
-            ...next.value?.dataCollectionProps,
-          };
-        } else if (next.type === TYPES.LIFI_VALUE_BRIDGED) {
-          step.prompt = next.value?.prompt;
-          step["additionalData"] = {
-            usdValue: toCent(next.value),
-          };
-        } else if (next.type === TYPES.VERIFY_MARKETSFLARE_TRIAL || Object.values(APEIRON_TYPES).includes(next.type)) {
-          step.prompt = next.value;
-        }
-        return [...acc, step];
-      }, []),
-    };
+    const body = constructRequestBody({ questSettings, steps, status, activeOrgId: activeOrg?.id });
     try {
       await questValidator(body);
-      if (!questSettings.isActive && !isSaving) {
-        return setIsSaving(true);
-      }
-
-      handleMutation({ body });
-
-      const hasMediaToUpload = steps.some(
-        (step) => step.mediaUploads?.filter((media) => media instanceof File).length > 0
-      );
-
-      const hasMediaToRemove = Object.values(removedMediaSlugs).flat().length > 0;
-      if (hasMediaToUpload || hasMediaToRemove) {
-        setSnackbarAlertMessage("Wrapping up with your media. Please keep this window open");
-        setSnackbarAlertAutoHideDuration(2000);
-        setSnackbarAlertOpen(true);
-      }
+      await processSave({
+        body,
+        handleMutation,
+        setIsSaving,
+        steps,
+        removedMediaSlugs,
+        setSnackbarAlertOpen,
+        setSnackbarAlertMessage,
+        setSnackbarAlertAutoHideDuration,
+        isActive: questSettings.isActive,
+        isSaving,
+      });
     } catch (err) {
-      const errors: any = {};
-      if (err instanceof ValidationError) {
-        err?.inner?.forEach((error) => {
-          const path = getPathArray(error.path);
-          set(errors, path, error.message);
-        });
-        // this is a hacky way to scroll to the title
-
-        if (errors?.title) {
-          window.scrollTo({
-            top: 0,
-            behavior: "smooth",
-          });
-        } else {
-          if (typeof errors?.steps === "string") {
-            refs?.current[0].scrollIntoView({ behavior: "smooth", block: "center" });
-            setErrors(errors);
-            setIsSaving(false);
-            return;
-          }
-          const stepsFirstErrorIndex = errors?.steps?.findIndex((err) => !!err);
-          if (stepsFirstErrorIndex !== -1) {
-            refs?.current[stepsFirstErrorIndex + 1]?.scrollIntoView({ behavior: "smooth", block: "center" });
-          }
-        }
-        setErrors(errors);
-        setIsSaving(false);
-      } else console.log(err, "Error outside of validation service");
+      handleSaveError({ err, setErrors, refs, setIsSaving });
     }
   };
 
   useMemo(() => setRefValue({ handleSave }), [setRefValue, handleSave]);
 
-  const handleRewardsToggle = () => {
-    setIsRewardModalOpen((prev) => !prev);
-    postRewardsToggle();
-  };
-
-  const onRewardsChange = (rewards) => {
-    setQuestSettings((prev) => ({
-      ...prev,
-      rewards,
-    }));
-  };
   const hasReferralStep = steps?.some((step) => step.type === TYPES.REFERRAL);
   return (
     <>
@@ -505,7 +280,7 @@ const CreateTemplate = ({
               }}
               renderHeader={null}
               renderBody={() => (
-                <RewardComponent
+                <QuestRewardComponent
                   rewards={questSettings.rewards}
                   setQuestSettings={setQuestSettings}
                   hasReferralStep={hasReferralStep}
