+++ wondrous-bot-admin/src/components/Pricing/PricingOptionsListItem.tsx
@@ -7,6 +7,7 @@ import { PricingOptionsListItemInnerWrapper, PricingOptionsListItemWrapper } fro
 import { useSubscription } from "utils/hooks";
 import { format } from "date-fns";
 import { useMe } from "components/Auth";
+import { getPlan } from "utils/common";
 
 export enum PricingOptionsTitle {
   Basic = "Basic",
@@ -56,12 +57,6 @@ const formatCurrency = (amount: number) =>
     maximumFractionDigits: 0,
   });
 
-export const getPlan = (plan) => {
-  if (!plan) return PricingOptionsTitle.Basic;
-  if (plan === "hobby" || plan === "hobby_annual") return PricingOptionsTitle.Hobby;
-  if (plan === "premium" || plan === "premium_annual") return PricingOptionsTitle.Premium;
-  if (plan === "ecosystem" || plan === "ecoosystem_annual") return PricingOptionsTitle.Ecosystem;
-};
 const STRIPE_MANAGE_SUBSCRIPTION_LINK = import.meta.env.VITE_PRODUCTION
   ? "https://billing.stripe.com/p/login/fZefYZfFDdyk6NG8ww"
   : "https://billing.stripe.com/p/login/test_3csbKGfr73hIg2QdQQ";
