+++ wondrous-bot-admin/src/components/Rewards/constants.ts
@@ -1,0 +1,13 @@
+export const REWARD_TYPES = [
+  { label: "ERC20", value: "erc20" },
+  { label: "ERC721", value: "erc721" },
+  { label: "ERC1155", value: "erc1155" },
+];
+
+export const PAYMENT_OPTIONS = {
+  DISCORD_ROLE: "discord_role",
+  TOKEN: "token",
+  POAP: "poap",
+  COMMUNITY_BADGE: "COMMUNITY_BADGE",
+  CMTY_STORE_ITEM: "cmty_store_item",
+};
