+++ wondrous-bot-admin/src/components/ViewQuestResults/ViewRewards.tsx
@@ -1,7 +1,7 @@
 import { Grid, Typography } from "@mui/material";
-import { PAYMENT_OPTIONS } from "components/CreateTemplate/RewardUtils";
 import { Label } from "components/CreateTemplate/styles";
 import { DiscordRoleIcon, NFTIcon, PointsIcon, StoreItemRewardIcon, TokensIcon } from "components/Icons/Rewards";
+import { PAYMENT_OPTIONS } from "components/Rewards/constants";
 import { constructRewards } from "utils/common";
 
 const selectReward = ({ type, value }) => {
@@ -25,7 +25,7 @@ const selectReward = ({ type, value }) => {
     [PAYMENT_OPTIONS.CMTY_STORE_ITEM]: {
       text: value,
       Icon: StoreItemRewardIcon,
-    }
+    },
   };
   return rewardProps[type] || { text: `${value} ${type}`, Icon: TokensIcon };
 };
